// phoenix-common-react(1)/src/core/featureToggles/loginClientBean.types.ts
import type { TaxInfo } from '../utils/utils.types';

export interface MinifiedLoginClientBean {
  officeTimezone: string;
  username: string;
  ldapUsername: string;
  userFullname: string;
  dataSourceName: string;
  schema: string;
  email: string;
  timeZone: string;
  office: string;
  company: string;
  localCurrency: string;
  userRegionId: number;
  countryCode: string;
  countryName: string;
  userCompanyName: string;
  userSchemaID: number;
  userOfficeID: number;
  userRoleID: number;
  userRole: string;
  userAlternateOffice: string;
  userId: number;
  siteId: number;
  officeId:number;
  locale: string;
}

export interface LoginClientBeanRaw {
  username: string;
  ldapUser: string;
  userName: string;
  userId: number;
  email: string;
  emailAlias?: string;
  office: string;
  officeId: number;
  officeTimezone: string;
  alternateCityName?: string;
  ldapOfficeCode?: string;
  userAlternateOffice?: string;
  userMainOffice: string;
  userMainOfficeCountry?: string;
  officeRegionCode: string;
  country: string;
  countryCode: string;
  latitude?: string;
  longitude?: string;
  schema: string;
  siteId: number;
  company: string;
  companyId: string;
  companyCode: string;
  companyName: string;
  companyGroupName: string;
  companyWebSite?: string;
  localCurrency: string;
  dateTimeFormat: string;
  locale: string;
  officeLocale?: string;
  timeZone?: string;
  userRole: string;
  userRoleID: number;
  userRegionId?: number;
  isToggleCustomer?: boolean;
  isLocalUser?: boolean;
  passwordExpired?: number;
  operationMode?: string;
  phoenixEnvironment: string;
  phoenixApiUrl?: string;
  oAuthToken?: string;
  userMainCompany: string;
  officeSettingMap: Record<string, string[]>;
  userSettingMap: Record<string, string[]>;
  programSettingMap: Record<string, Record<string, string[]>>;
  locationSettingMap: Record<string, Record<string, string[]>>;
  companySettingMap: Record<string, string>;
  countryMap: Record<string, string>;
  currencyDecimalMap?: Record<string, number>;
  taxSettingMap?: Record<string, TaxInfo>;
  usOfficeMap?: Record<string, string[]>;
  config_environment_customer: string;
  [key: string]: unknown;
}